refactor(validate-linux-packages): extract validate_packages.py into focused modules
Improve cohesion by refactoring the monolithic validate_packages.py (734 lines)
into 5 focused modules, reducing it to 377 lines of orchestration logic.

New modules created:

- validate_formatters.py (51 lines): Output trimming and stderr extraction
  - _trim_output() - Multi-line diagnostic formatting
  - _trim_output_single_line() - Single-line logging format
  - _extract_process_stderr() - Process error extraction

- validate_arch.py (70 lines): Architecture mapping and validation
  - acceptable_rpm_architectures() - RPM arch aliases
  - rpm_expected_architecture() - Canonical RPM arch
  - _should_skip_sandbox() - Cross-arch skip logic
  - _host_architectures() - Platform detection

- validate_locators.py (49 lines): Package discovery
  - locate_deb() - Find Debian packages
  - locate_rpm() - Find RPM packages
  - ensure_subset() - Validate file presence

- validate_sandbox_diagnostics.py (176 lines): Diagnostic collection
  - _format_path_diagnostics() - Path inspection
  - _collect_environment_details() - Sandbox environment
  - _collect_host_path_details() - Host-side path info
  - _execute_diagnostic_command() - Command diagnostics

- validate_path_checks.py (151 lines): Path validation
  - _validate_paths_exist() - Existence checks
  - _validate_paths_executable() - Executability checks
  - _try_python_fallback() - Python os.access fallback

Changes to validate_packages.py:

- Removed 357 lines of extracted functions
- Added imports from new modules
- Maintained __all__ exports for backward compatibility
- Retained orchestration functions:
  - _exec_with_diagnostics()
  - _install_and_verify()
  - _validate_package()
  - _MetadataValidators
  - validate_deb_package()
  - validate_rpm_package()

Resolution of duplicate _trim_output():

The file had two _trim_output() functions with different signatures. The second
definition (line 385) was overriding the first (line 148). Investigation revealed
they serve different purposes:
- Multi-line variant (line 148): Diagnostic formatting with line limits
- Single-line variant (line 385): Logging with whitespace normalisation

Resolved by renaming the second to _trim_output_single_line() and updating
call sites at lines 419, 426.

Backward compatibility:

All public functions remain importable from validate_packages via re-exports.
All 635 tests pass without modification. Tests use dynamic module loading which
works seamlessly with the refactored structure.

Validation:

- make typecheck: PASSED (1 pre-existing warning unrelated to changes)
- make test: PASSED (635 passed, 86 skipped)
- make lint: PASSED
- make check-fmt: PASSED

Implements execplan: docs/execplans/refactor-validate-packages-modules.md
